缩放崩溃，修复

+++ GraphicEngine/src/d3dsamplebase.cpp
@@ -86,7 +86,7 @@ void D3DSampleBase::Present()
 
 float Graphic::D3DSampleBase::AspectRatio() const
 {
-	return static_cast<float>(clientWidth / clientHeight);
+	return static_cast<float>(clientWidth / (float)clientHeight);
 }
 
 void D3DSampleBase::Render()
